chore: rename fonts.js to typography.js
